ml_training/predict_spend/cloudrun_build_serving_image/requirements.txt
Flask==3.0.0
gunicorn==20.1.0
Werkzeug==3.0.1
google-cloud-build